Implement 'q_swap'
This commit introduces the 'q_swap' function to swap adjacent elements
in a doubly linked list. It uses the 'list_for_each_safe' macro to
safely traverse the list, which inherently checks that the current
element ('l1') is not the head. During each iteration, if 'l2' is not
equal to the list head, 'l1' is removed from its current position and
reinserted before 'l2'. After the swap, 'l2' is updated to 'l1->next' to
ensure that no element is processed more than once.

Change-Id: I1c8f22f44717bf4638786c41edbf432fda2f47b2